Added a few more classes for multiplayer execution. It is pretty bad now, but I am working on a more elegant and generic solution that can easily "dock" with out game agents.
I also altered a few parameters on the light and materials in order to find out why the color of the blocks changes based on what I load for the ball.

We now have a Pokeball! If anyone figures out what is the problem with the red color before me, fix and let me know :)